requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
jinja2==3.1.4
python-multipart==0.0.9
SQLAlchemy==2.0.31
pyodbc==5.0.1
python-dotenv==1.0.1
reportlab==4.2.2
